1.44.0.0: Release #2253

+++ source/res/scripts/client/gui/impl/lobby/crew/base_crew_view.py
@@ -1,6 +1,6 @@
 # Python bytecode 2.7 (decompiled from Python 2.7)
 # Embedded file name: scripts/client/gui/impl/lobby/crew/base_crew_view.py
-from typing import TYPE_CHECKING, Optional
+from typing import TYPE_CHECKING, Optional, Union
 from AccountCommands import LOCK_REASON
 from PlayerEvents import g_playerEvents
 from crew_sounds import CREW_SOUND_SPACE, CREW_SOUND_OVERLAY_SPACE
@@ -241,5 +241,5 @@ class BaseCrewView(BaseCrewSoundView, IGlobalListener):
         self.destroyWindow()
 
     def _onVehicleLockChanged(self, vehInvID, lockReason):
-        if lockReason[0] in (LOCK_REASON.PREBATTLE, LOCK_REASON.UNIT):
+        if lockReason and lockReason[0] in (LOCK_REASON.PREBATTLE, LOCK_REASON.UNIT):
             self._destroySubViews()
